requirements.txt
annotated-types==0.7.0
anyio==4.8.0
attrs==23.2.0
bcrypt==4.0.1
beautifulsoup4==4.12.3
blinker==1.8.2
bs4==0.0.2
cachetools==5.5.2
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.8
colorama==0.4.6
defusedxml==0.7.1
ecdsa==0.19.1
exceptiongroup==1.2.2
fastapi==0.115.11
Flask==3.0.3
fonttools==4.49.0
fpdf2==2.7.8
google-api-core==2.25.0rc1
google-auth==2.40.2
google-cloud-core==2.4.3
google-cloud-storage==3.1.0
google-crc32c==1.7.1
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
greenlet==3.1.1
gunicorn==23.0.0
h11==0.16.0
html5lib==1.1
idna==3.10
importlib_metadata==7.1.0
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
meson==1.4.1
numpy==1.26.4
ordered-set==4.1.0
outcome==1.3.0.post0
packaging==24.2
pandas==2.2.2
passlib==1.7.4
pillow==10.2.0
playwright==1.51.0
pluggy==1.4.0
proto-plus==1.26.1
protobuf==6.31.0
py==1.11.0
pyaes==1.6.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
pyee==12.1.1
pygame==2.5.2
PyLaTeX==1.4.2
pyodbc==5.2.0
PyQt5==5.15.9
PyQt5-Qt5==5.15.2
PyQt5-sip==12.13.0
PySocks==1.7.1
pyTelegramBotAPI==4.16.1
pytest==8.0.2
python-dateutil==2.9.0.post0
python-jose==3.4.0
python-multipart==0.0.20
pytz==2024.1
requests==2.31.0
rsa==4.9
schedule==1.2.1
selenium==4.20.0
setuptools==70.0.0
setuptools-metadata==0.1.5
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.5
starlette==0.46.0
stripe==12.1.0
telebot==0.0.5
Telethon==1.37.0
trio==0.25.0
trio-websocket==0.11.1
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.1
uvicorn==0.34.2
webencodings==0.5.1
Werkzeug==3.0.3
wheel==0.43.0
wsproto==1.2.0
zipp==3.19.2
